requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
agno>=0.1.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
pydantic>=2.5.0
fastapi-cors>=0.0.6
python-multipart>=0.0.6